Fixed shading mode and changed to checkbox
Grid cells now shade depending on the color of the cell.
This is done by parsing the rgba values of the background color of a cell
and increasing the alpha value by 10% each time the cell is entered.
Shading mode button changed from radio button to checkbox to allow to start drawing by shading.